Removed incredibly bad use of C-style casting combined with RTTI static casting such as:
mMotionStar.setBirdRate((double)(static_cast<int>(c->getProperty("brate"))));


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@2042 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/gadgeteer/gadget/Devices/Ascension/MotionStar/MotionStar.cpp
@@ -107,13 +107,13 @@ vjMotionStar::config (vjConfigChunk *c) {
                               << vjDEBUG_FLUSH;
 
    // set mMotionStar with the config info.
-   mMotionStar.setHemisphere((unsigned int)(static_cast<int> (c->getProperty("hemisphere"))));
-   mMotionStar.setBirdFormat((unsigned int)(static_cast<int> (c->getProperty("bformat"))));
-   mMotionStar.setNumBirds((unsigned int)(static_cast<int> (c->getProperty("num"))));
-   mMotionStar.setBirdRate((double)(static_cast<int>(c->getProperty("brate"))));
-   mMotionStar.setRunMode((static_cast<int> (c->getProperty("mode"))));
+   mMotionStar.setHemisphere((unsigned int) static_cast<int>(c->getProperty("hemisphere")));
+   mMotionStar.setBirdFormat((unsigned int) static_cast<int>(c->getProperty("bformat")));
+   mMotionStar.setNumBirds((unsigned int) static_cast<int>(c->getProperty("num")));
+   mMotionStar.setBirdRate(static_cast<double>(c->getProperty("brate")));
+   mMotionStar.setRunMode(static_cast<int>(c->getProperty("mode")));
    mMotionStar.setAddress(static_cast<std::string>(c->getProperty("address")).c_str());
-//   mMotionStar.setReportRate( (static_cast<unsigned char>(c->getProperty("rrate"))) ); 
+//   mMotionStar.setReportRate(static_cast<unsigned char>(c->getProperty("rrate"))); 
  
    return true;
 }
